+++ modules/error_manager/builtin/cpp/Gateway.cpp
@@ -29,7 +29,7 @@
 #include "lastwarnBuiltin.hpp"
 #include "warningBuiltin.hpp"
 #include "getLastReportBuiltin.hpp"
-#include "mexception_extractionBuiltin.hpp"
+#include "MException_extractionBuiltin.hpp"
 //=============================================================================
 using namespace Nelson;
 //=============================================================================
@@ -45,7 +45,7 @@ static const nlsGateway gateway[] = {
         CPP_BUILTIN_WITH_EVALUATOR },
     { "getLastReport", (void*)Nelson::ErrorManagerGateway::getLastReportBuiltin, 1, 0,
         CPP_BUILTIN_WITH_EVALUATOR },
-    { "mexception_extraction", (void*)Nelson::ErrorManagerGateway::mexception_extractionBuiltin, 0,
+    { "MException_extraction", (void*)Nelson::ErrorManagerGateway::MException_extractionBuiltin, 0,
         1, CPP_BUILTIN },
 };
 //=============================================================================